Add matrix rain animation, and implement _sbrk syscall.
It seems that libc insists on initializing the heap to make use of
srand()/rand(). So we might as well include an _sbrk call. This is
implemented in the animation rather than via an ecall, since the
BIOS isn't really aware of the .bss layout of the animation to do
heap management correctly.